% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_v01.R
\name{build_url}
\alias{build_url}
\title{Build complex URL}
\usage{
build_url(base_url, ticker, url_params = NULL)
}
\arguments{
\item{base_url}{string of the base URL}

\item{ticker}{string corresponding to a ticker of interest}

\item{url_params}{list of additional parameters}
}
\value{
string with the URL for querying yahoo finance
}
\description{
Build URL to query Yahoo! Finance
}
\examples{
build_url(base_url='http://www.google.com', ticker='AAPL')

}
\author{
Damiano Fantini <damiano.fantini@gmail.com>
}
