% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_01.R
\docType{package}
\name{dffin82-package}
\alias{dffin82-package}
\title{Manage Financial Data from Yahoo! using R.}
\description{
Yahoo! Finance is a part of the Yahoo! network and is an online
service providing financial news, data and commentary including 
stock quotes, press releases, financial reports, and other original content. 
The dffin82 library is a platform providing a series of functions and tools 
for downloading, managing, aggregating, and analyzing stock data from 
Yahoo! Finance.
}
\references{
\enumerate{
  \item{\strong{Yahoo! Finance Website}: \url{https://finance.yahoo.com/} }
  \item{\strong{GitHub Page}: \url{https://en.wikipedia.org/wiki/GitHub} }
 }
}
