% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_v01.R
\name{fix_url}
\alias{fix_url}
\title{Fix URL string}
\usage{
fix_url(url)
}
\arguments{
\item{url}{string of a URL to fix}
}
\value{
string with the corrected URL
}
\description{
Fix the URL string
}
\examples{
fix_url('http://www.google.com')

}
\author{
Damiano Fantini <damiano.fantini@gmail.com>
}
