% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_v01.R
\name{vector_shift}
\alias{vector_shift}
\title{Shift Vector}
\usage{
vector_shift(x, step = 1)
}
\arguments{
\item{x}{input data, typically a numeric or character vector}

\item{step}{integer, number of positions}
}
\value{
vector including the shifted values
}
\description{
Shift Values of a Vector
}
\examples{
vector_shift(1:5)

}
\author{
Damiano Fantini <damiano.fantini@gmail.com>
}
